Phase K: fix self-aliasing sprintf in sql_query + 0x0071/0x0072 round-trip
The mysqlplus.cpp legacy SQL builder uses sprintf_s(buf, sz, "%s, X", buf, ...)
to append into m_Values/m_Fields. On Win32 the MSVC runtime's sprintf_s
tolerates the overlap; on Linux the same call routes through vsnprintf,
which is UB when the destination overlaps a vararg, and glibc silently
truncates every multi-field INSERT to its last column. Concretely,
CreateCharacter emitted INSERT INTO "avatar_info" ("trade") VALUES ('0')
and Postgres rejected it on the NOT-NULL avatar_id constraint, so the
GlobalCreateCharacter exchange completed at the wire layer but the new
avatar never persisted.

Fix: introduce AppendValue() that formats into a small scratch buffer
then strncat()s onto the destination — no aliasing — and route every
AddData/AddField variant through it. Also rewrite the string AddData
to format into m_Values directly via offset tracking so it isn't capped
at the 256B scratch.

Coverage adds: 0x0072 GLOBAL_CREATE_CHARACTER round-trip (cli_test03
creates Testavus the Terran Warrior in slot 0 sector 10151, then deletes
to clean up) plus the prior 0x0071 DELETE_CHARACTER round-trip. Both
exercise the Phase K ColorInfo / ShipData / GlobalCreateCharacter
int32_t wire-size migration and the GlobalAvatarListCodec AvatarData
offset fix (race=46 prof=50 gender=54 mood=58, not 48/52/56/60 — the
codec had stale comments assuming 2-byte alignment padding after
filler1+avatar_version, but ATTRIB_PACKED disables that).

Coverage floor: 11.

+++ common/include/net7/PacketStructures.h
@@ -177,9 +177,20 @@ struct AvatarInfo
 
 struct ColorInfo
 {
+    // Phase K: int32_t (was `long`). Win32 packed = 12 + 1 + 4 = 17 bytes;
+    // Linux with 8-byte long = 21 bytes. Embedded 8× in ShipData which is
+    // embedded in GlobalCreateCharacter — pre-migration GCC bloated
+    // ShipData to 226 bytes and GlobalCreateCharacter to 571 bytes (vs
+    // canonical Win32 539). Sent verbatim via sizeof(GlobalCreateCharacter)
+    // in UDPClient_linux.cpp::CreateCharacter so any Linux client paying
+    // attention to Win32 wire size would corrupt the server's parse.
+    // Server-side `metal` consumers (server/src/PlayerMisc.cpp,
+    // AccountManager.cpp) all assign from a local `int metal`, so
+    // narrowing long→int32_t is value-preserving for every existing
+    // call site.
     float   HSV[3];
     char    flat;
-    long    metal;
+    int32_t metal;
 } ATTRIB_PACKED;  // 17 bytes
 
 struct ShipData
